requirements.txt
absl-py==0.9.0
anaconda-clean==1.0
anaconda-client==1.7.2
anaconda-navigator==1.9.12
appdirs==1.4.3
applaunchservices==0.2.1
appnope==0.1.0
arrow==0.15.5
asgiref==3.2.3
asn1crypto==1.3.0
astor==0.8.1
attrs==19.3.0
backcall==0.1.0
backports.functools-lru-cache==1.6.1
backports.tempfile==1.0
backports.weakref==1.0.post1
beautifulsoup4==4.8.2
bleach==3.1.0
cachetools==4.1.0
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
click==7.1.1
clyent==1.2.2
conda==4.8.3
conda-build==3.18.11
conda-package-handling==1.6.0
conda-verify==3.4.2
cryptography==2.8
cycler==0.10.0
Cython==0.29.16
decorator==4.4.2
defusedxml==0.6.0
diff-match-patch==20181111
dill==0.3.1.1
distlib==0.3.0
Django==3.0.5
django-cors-headers==3.2.1
django-hitcount==1.3.1
django-taggit==1.2.0
django-taggit-serializer==0.1.7
djangorestframework==3.11.0
dlx==1.0.4
docplex==2.13.184
entrypoints==0.3
fastdtw==0.3.4
fastjsonschema==2.14.4
filelock==3.0.12
future==0.18.2
gast==0.2.2
glob2==0.7
google-auth==1.13.1
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
grpcio==1.28.1
h5py==2.10.0
helpdev==0.6.10
idna==2.9
imageio==2.8.0
importlib-metadata==1.5.0
inflection==0.4.0
ipykernel==5.1.4
ipython==7.13.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
jedi==0.16.0
Jinja2==2.11.1
joblib==0.14.1
json5==0.9.4
jsonschema==3.2.0
jupyter-client==6.1.2
jupyter-contrib-core==0.3.3
jupyter-contrib-nbextensions==0.5.1
jupyter-core==4.6.3
jupyter-highlight-selected-word==0.2.0
jupyter-latex-envs==1.4.6
jupyter-nbextensions-configurator==0.4.1
jupyterlab==1.2.6
jupyterlab-server==1.1.0
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.2.0
libarchive-c==2.8
lief==0.9.0
lxml==4.5.0
Markdown==3.2.1
MarkupSafe==1.1.1
marshmallow==3.5.1
marshmallow-polyfield==5.9
matplotlib==3.2.1
mistune==0.8.4
more-itertools==8.2.0
mpmath==1.1.0
navigator-updater==0.2.1
nbconvert==5.6.1
nbformat==5.0.4
nest-asyncio==1.3.2
networkx==2.4
notebook==6.0.3
ntlm-auth==1.4.0
numpy==1.18.2
oauthlib==3.1.0
olefile==0.46
opt-einsum==3.2.0
pandas==1.0.3
pandocfilters==1.4.2
parso==0.6.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.0.0
pkginfo==1.5.0.1
ply==3.11
prometheus-client==0.7.1
prompt-toolkit==3.0.4
protobuf==3.11.3
psutil==5.7.0
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pybind11==2.5.0
pycosat==0.6.3
pycparser==2.20
Pygments==2.6.1
pyobjc==6.1
pyobjc-core==6.1
pyobjc-framework-Accounts==6.1
pyobjc-framework-AddressBook==6.1
pyobjc-framework-AdSupport==6.1
pyobjc-framework-AppleScriptKit==6.1
pyobjc-framework-AppleScriptObjC==6.1
pyobjc-framework-ApplicationServices==6.1
pyobjc-framework-AuthenticationServices==6.1
pyobjc-framework-Automator==6.1
pyobjc-framework-AVFoundation==6.1
pyobjc-framework-AVKit==6.1
pyobjc-framework-BusinessChat==6.1
pyobjc-framework-CalendarStore==6.1
pyobjc-framework-CFNetwork==6.1
pyobjc-framework-CloudKit==6.1
pyobjc-framework-Cocoa==6.1
pyobjc-framework-Collaboration==6.1
pyobjc-framework-ColorSync==6.1
pyobjc-framework-Contacts==6.1
pyobjc-framework-ContactsUI==6.1
pyobjc-framework-CoreAudio==6.1
pyobjc-framework-CoreAudioKit==6.1
pyobjc-framework-CoreBluetooth==6.1
pyobjc-framework-CoreData==6.1
pyobjc-framework-CoreHaptics==6.1
pyobjc-framework-CoreLocation==6.1
pyobjc-framework-CoreMedia==6.1
pyobjc-framework-CoreMediaIO==6.1
pyobjc-framework-CoreML==6.1
pyobjc-framework-CoreMotion==6.1
pyobjc-framework-CoreServices==6.1
pyobjc-framework-CoreSpotlight==6.1
pyobjc-framework-CoreText==6.1
pyobjc-framework-CoreWLAN==6.1
pyobjc-framework-CryptoTokenKit==6.1
pyobjc-framework-DeviceCheck==6.1
pyobjc-framework-DictionaryServices==6.1
pyobjc-framework-DiscRecording==6.1
pyobjc-framework-DiscRecordingUI==6.1
pyobjc-framework-DiskArbitration==6.1
pyobjc-framework-DVDPlayback==6.1
pyobjc-framework-EventKit==6.1
pyobjc-framework-ExceptionHandling==6.1
pyobjc-framework-ExecutionPolicy==6.1
pyobjc-framework-ExternalAccessory==6.1
pyobjc-framework-FileProvider==6.1
pyobjc-framework-FileProviderUI==6.1
pyobjc-framework-FinderSync==6.1
pyobjc-framework-FSEvents==6.1
pyobjc-framework-GameCenter==6.1
pyobjc-framework-GameController==6.1
pyobjc-framework-GameKit==6.1
pyobjc-framework-GameplayKit==6.1
pyobjc-framework-ImageCaptureCore==6.1
pyobjc-framework-IMServicePlugIn==6.1
pyobjc-framework-InputMethodKit==6.1
pyobjc-framework-InstallerPlugins==6.1
pyobjc-framework-InstantMessage==6.1
pyobjc-framework-Intents==6.1
pyobjc-framework-IOSurface==6.1
pyobjc-framework-iTunesLibrary==6.1
pyobjc-framework-LatentSemanticMapping==6.1
pyobjc-framework-LaunchServices==6.1
pyobjc-framework-libdispatch==6.1
pyobjc-framework-LinkPresentation==6.1
pyobjc-framework-LocalAuthentication==6.1
pyobjc-framework-MapKit==6.1
pyobjc-framework-MediaAccessibility==6.1
pyobjc-framework-MediaLibrary==6.1
pyobjc-framework-MediaPlayer==6.1
pyobjc-framework-MediaToolbox==6.1
pyobjc-framework-MetalKit==6.1
pyobjc-framework-ModelIO==6.1
pyobjc-framework-MultipeerConnectivity==6.1
pyobjc-framework-NaturalLanguage==6.1
pyobjc-framework-NetFS==6.1
pyobjc-framework-Network==6.1
pyobjc-framework-NetworkExtension==6.1
pyobjc-framework-NotificationCenter==6.1
pyobjc-framework-OpenDirectory==6.1
pyobjc-framework-OSAKit==6.1
pyobjc-framework-OSLog==6.1
pyobjc-framework-PencilKit==6.1
pyobjc-framework-Photos==6.1
pyobjc-framework-PhotosUI==6.1
pyobjc-framework-PreferencePanes==6.1
pyobjc-framework-PubSub==6.1
pyobjc-framework-PushKit==6.1
pyobjc-framework-Quartz==6.1
pyobjc-framework-QuickLookThumbnailing==6.1
pyobjc-framework-SafariServices==6.1
pyobjc-framework-SceneKit==6.1
pyobjc-framework-ScreenSaver==6.1
pyobjc-framework-ScriptingBridge==6.1
pyobjc-framework-SearchKit==6.1
pyobjc-framework-Security==6.1
pyobjc-framework-SecurityFoundation==6.1
pyobjc-framework-SecurityInterface==6.1
pyobjc-framework-ServiceManagement==6.1
pyobjc-framework-Social==6.1
pyobjc-framework-SoundAnalysis==6.1
pyobjc-framework-Speech==6.1
pyobjc-framework-SpriteKit==6.1
pyobjc-framework-StoreKit==6.1
pyobjc-framework-SyncServices==6.1
pyobjc-framework-SystemConfiguration==6.1
pyobjc-framework-SystemExtensions==6.1
pyobjc-framework-UserNotifications==6.1
pyobjc-framework-VideoSubscriberAccount==6.1
pyobjc-framework-VideoToolbox==6.1
pyobjc-framework-Vision==6.1
pyobjc-framework-WebKit==6.1
pyOpenSSL==19.1.0
pyparsing==2.4.7
pyrsistent==0.16.0
pyscf==1.7.1.post1
PySocks==1.7.1
python-constraint==1.4.0
python-dateutil==2.8.1
pytz==2019.3
PyYAML==5.3.1
pyzmq==18.1.1
QDarkStyle==2.8
qiskit==0.18.0
qiskit-aer==0.5.0
qiskit-aqua==0.6.5
qiskit-ibmq-provider==0.6.0
qiskit-ignis==0.3.0
qiskit-terra==0.13.0
QtPy==1.9.0
Quandl==3.5.0
requests==2.23.0
requests-ntlm==1.1.0
requests-oauthlib==1.3.0
retworkx==0.3.3
rsa==4.0
ruamel-yaml==0.15.87
scikit-learn==0.22.2.post1
scipy==1.4.1
Send2Trash==1.5.0
six==1.14.0
soupsieve==2.0
sqlparse==0.3.1
sympy==1.5.1
tensorboard==2.1.1
tensorflow==2.1.0
tensorflow-estimator==2.1.0
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
tornado==6.0.4
tqdm==4.44.1
traitlets==4.3.3
urllib3==1.25.8
virtualenv==20.0.18
wcwidth==0.1.9
webencodings==0.5.1
websockets==7.0
Werkzeug==1.0.1
widgetsnbextension==3.5.1
wrapt==1.12.1
xmltodict==0.12.0
zipp==2.2.0

git+https://github.com/mateusz800/django-hitcount